Can now run multiple times

+++ 160_gitlab_ci/000_rollout/bootstrap.sh
@@ -118,17 +118,19 @@ git config --global user.name "seat"
 git config --global user.email "seat@seat${SEAT_INDEX}.inmylab.de"
 git config --global credential.helper store
 echo "https://seat:${SEAT_PASS}@gitlab.seat${SEAT_INDEX}.inmylab.de" >"${HOME}/.git-credentials"
-(
-    mkdir -p /tmp/demo
-    cd /tmp/demo
-    git clone "https://gitlab.seat${SEAT_INDEX}.inmylab.de/seat/demo" .
-    git checkout --orphan main
-    git rm -rf .
-    git commit --allow-empty -m "root commit"
-    git push origin main
-    cd /tmp
-    rm -rf demo
-)
+if ! git show-ref --quiet refs/heads/main; then
+    (
+        mkdir -p /tmp/demo
+        cd /tmp/demo
+        git clone "https://gitlab.seat${SEAT_INDEX}.inmylab.de/seat/demo" .
+        git checkout --orphan main
+        git rm -rf .
+        git commit --allow-empty -m "root commit"
+        git push origin main
+        cd /tmp
+        rm -rf demo
+    )
+fi
 docker compose exec -T gitlab \
     curl \
         --url "http://localhost/api/v4/projects/seat%2fdemo" \
